+++ ask-sdk-express-adapter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+# 2.1.0 (2020-05-07)
+
+This release contains the following changes : 
+
+- Add certificate chain validation by using node-forge pacakge.
+
 # 2.0.1 (2020-01-22)
 
 This release contains the following changes :